fix: extract 403 and 404 page
